# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi API-TA Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "asae_marcial@hotmail.es"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/getPostSeguidos": {
            "get": {
                "tags": [
                    "POSTS"
                ],
                "summary": "Obtiene las publicaciones de los usuarios seguidos.",
                "operationId": "getPostSeguidos",
                "responses": {
                    "200": {
                        "description": "Solicitud exitosa"
                    },
                    "400": {
                        "description": "Error de solicitud"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/misPosts": {
            "get": {
                "tags": [
                    "POSTS"
                ],
                "summary": "Obtiene las publicaciones de los usuarios logeado.",
                "operationId": "misPosts",
                "responses": {
                    "200": {
                        "description": "Solicitud exitosa"
                    },
                    "400": {
                        "description": "Error de solicitud"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/createPost": {
            "post": {
                "tags": [
                    "POSTS"
                ],
                "summary": "Crea un publicación.",
                "operationId": "createPost",
                "parameters": [
                    {
                        "name": "descripcion",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ruta",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud exitosa"
                    },
                    "400": {
                        "description": "Error de solicitud"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/{idPost}/editPost": {
            "put": {
                "tags": [
                    "POSTS"
                ],
                "summary": "Edita un publicación.",
                "operationId": "editPost",
                "parameters": [
                    {
                        "name": "idPost",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud exitosa"
                    },
                    "400": {
                        "description": "Error de solicitud"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/{idPost}/deletePost": {
            "delete": {
                "tags": [
                    "POSTS"
                ],
                "summary": "Elimina una publicación.",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "idPost",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud exitosa"
                    },
                    "400": {
                        "description": "Error de solicitud"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "CUENTA"
                ],
                "summary": "Inicia sesion.",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error de solicitud"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "CUENTA"
                ],
                "summary": "Crea una cuenta.",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellidos",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "info",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud exitosa"
                    },
                    "400": {
                        "description": "Error de solicitud"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Inicie sesión con correo electrónico y contraseña para obtener el token de autenticación",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}